refactor(api): fix typo in usecases name
